Reference

Base.getindex — Method

Get symbolic representation for function index object

Base.isconst — Method

isconst(x::Constant)

True if the symbol value cannot be modified within an Operator (and thus its value is provided by the user directly from Python-land), False otherwise.

Base.ndims — Method

ndims(x::DiscreteFunction)

Return the number of dimensions corresponding to the discrete function x.

Base.size — Method

size(x::DiscreteFunction)

Return the shape of the grid for the discrete function x.

Devito.Max — Function

Max(args...)

Can be used in a Devito.Eq to return the minimum of a collection of arguments Example:

    eqmax = Eq(f,Max(g,1))

Is equivalent to f = Max(g,1) for Devito functions f,g

Devito.Min — Function

Min(args...)

Can be used in a Devito.Eq to return the minimum of a collection of arguments Example:

    eqmin = Eq(f,Min(g,1))

Is equivalent to f = Min(g,1) for Devito functions f,g

Devito.Mod — Method

Mod(x::AbstractDimension,y::Int)

Perform Modular division on a dimension

Devito.apply — Method

apply( operator::Operator; kwargs...)

Execute the Devito operator, Operator.

See: https://www.devitoproject.org/devito/operator.html?highlight=apply#devito.operator.operator.Operator.apply

Note that this returns a summary::Dict of the action of applying the operator. This contains information such as the number of floating point operations executed per second.

Devito.backward — Method

backward(x::TimeFunction)

Returns the symbol for the time-backward state of the TimeFunction.

See: https://www.devitoproject.org/devito/timefunction.html?highlight=forward#devito.types.TimeFunction.backward

Devito.ccode — Method

```
ccode(x::Operator; filename="")
```

Print the ccode associated with a devito operator. If filename is provided, writes ccode to disk using that filename

Devito.configuration! — Method

configuration!(key, value)

Configure Devito. Examples include

configuration!("log-level", "DEBUG")
configuration!("language", "openmp")
configuration!("mpi", false)

Devito.data — Method

data(x::Constant{T})

Returns value(x::Constnat{T}). See value documentation for more information.

Devito.data — Method

data(x::DiscreteFunction)

Return the data associated with the grid that corresponds to the discrete function x. This is the portion of the grid that excludes the halo. In the case of non-MPI Devito, this returns an array of type DevitoArray. In the case of the MPI Devito, this returns an array of type DevitoMPIArray.

The data can be converted to an Array via convert(Array, data(x)). In the case where data(x)::DevitoMPIArray, this also collects the data onto MPI rank 0.

Devito.data_allocated — Method

data_allocated(x::DiscreteFunction)

Return the data associated with the grid that corresponds to the discrete function x. This is the portion of the grid that includes the inner halo and includes the outer halo. We expect this to be equivalent to data_with_inhalo.

The data can be converted to an Array via convert(Array, data(x)). In the case where data(x)::DevitoMPIArray, this also collects the data onto MPI rank 0.

Devito.data_with_halo — Method

data_with_halo(x::DiscreteFunction)

Return the data associated with the grid that corresponds to the discrete function x. This is the portion of the grid that excludes the inner halo and includes the outer halo. In the case of non-MPI Devito, this returns an array of type DevitoArray. In the case of the MPI Devito, this returns an array of type DevitoMPIArray.

The data can be converted to an Array via convert(Array, data(x)). In the case where data(x)::DevitoMPIArray, this also collects the data onto MPI rank 0.

Devito.data_with_inhalo — Method

data_with_inhalo(x::DiscreteFunction)

Return the data associated with the grid that corresponds to the discrete function x. This is the portion of the grid that includes the inner halo and includes the outer halo. In the case of non-MPI Devito, this returns an array of type DevitoArray. In the case of the MPI Devito, this returns an array of type DevitoMPIArray.

The data can be converted to an Array via convert(Array, data(x)). In the case where data(x)::DevitoMPIArray, this also collects the data onto MPI rank 0.

Devito.dimensions — Method

dimensions(x::Union{Grid,DiscreteFunction})

Returns a tuple with the dimensions associated with the Devito grid.

Devito.dt — Function

dt(f::TimeFunction, args...; kwargs...)

Returns the symbol for the first time derivative of a time function

Devito.dt2 — Function

dt2(f::TimeFunction, args...; kwargs...)

Returns the symbol for the second time derivative of a time function

Devito.dx — Function

dx(f::DiscreteFunction, args...; kwargs...)

Returns the symbol for the first derivative with respect to x.

Devito.dxl — Function

dxl(f::DiscreteFunction, args...; kwargs...)

Returns the symbol for the first backward one-sided derivative with respect to x.

Devito.dxr — Function

dxr(f::DiscreteFunction, args...; kwargs...)

Returns the symbol for the first forward one-sided derivative with respect to x.

Devito.dy — Function

dy(f::DiscreteFunction, args...; kwargs...)

Returns the symbol for the first derivative with respect to y.

Devito.dyl — Function

dyl(f::DiscreteFunction, args...; kwargs...)

Returns the symbol for the first backward one-sided derivative with respect to y.

Devito.dyr — Function

dyr(f::DiscreteFunction, args...; kwargs...)

Returns the symbol for the first forward one-sided derivative with respect to y.

Devito.dz — Function

dzr(f::DiscreteFunction, args...; kwargs...)

Returns the symbol for the first forward one-sided derivative with respect to z.

Devito.dzl — Function

dzl(f::DiscreteFunction, args...; kwargs...)

Returns the symbol for the first backward one-sided derivative with respect to z.

Devito.forward — Method

forward(x::TimeFunction)

Returns the symbol for the time-forward state of the TimeFunction.

See: https://www.devitoproject.org/devito/timefunction.html?highlight=forward#devito.types.TimeFunction.forward

Devito.grid — Method

grid(f::DiscreteFunction)

Return the grid corresponding to the discrete function f.

Devito.halo — Method

halo(x::DiscreteFunction)

Return the Devito "outer" halo size corresponding to the discrete function f.

Devito.inhalo — Method

inhalo(x::DiscreteFunction)

Return the Devito "inner" halo size used for domain decomposition, and corresponding to the discrete function f.

Devito.inject — Method

inject(x::SparseTimeFunction; kwargs...)

Generate equations injecting an arbitrary expression into a field.

See: Generate equations injecting an arbitrary expression into a field.

Example

x = SpaceDimension(name="x", spacing=Constant(name="h_x", value=5.0))
z = SpaceDimension(name="z", spacing=Constant(name="h_z", value=5.0))
grid = Grid(
    dimensions = (x,z),
    shape = (251,501), # assume x is first, z is second (i.e. z is fast in python)
    origin = (0.0,0.0),
    extent = (1250.0,2500.0),
    dtype = Float32)

time_range = 0.0f0:0.5f0:1000.0f0
src = SparseTimeFunction(name="src", grid=grid, npoint=1, nt=length(time_range))
src_term = inject(src; field=forward(p), expr=2*src)

Devito.interior — Method

interior(x::grid)

returns the interior subdomain of a Devito grid

Devito.interpolate — Method

interpolate(x::SparseTimeFunction; kwargs...)

Generate equations interpolating an arbitrary expression into self.

See: https://www.devitoproject.org/devito/sparsetimefunction.html#devito.types.SparseTimeFunction.interpolate

Example

x = SpaceDimension(name="x", spacing=Constant(name="h_x", value=5.0))
z = SpaceDimension(name="z", spacing=Constant(name="h_z", value=5.0))
grid = Grid(
    dimensions = (z,x),
    shape = (501,251), # assume z is first, x is second
    origin = (0.0,0.0),
    extent = (2500.0,1250.0),
    dtype = Float32)

p = TimeFunction(name="p", grid=grid, time_order=2, space_order=8)

time_range = 0.0f0:0.5f0:1000.0f0
nz,nx,δz,δx = size(grid)...,spacing(grid)...
rec = SparseTimeFunction(name="rec", grid=grid, npoint=nx, nt=length(time_range))
rec_coords = coordinates(rec)
rec_coords[1,:] .= 10.0
rec_coords[2,:] .= δx*(0:nx-1)

rec_term = interpolate(rec, expr=p)

Devito.is_Derived — Function

is_Derived(x::Dimension)

Returns true when dimension is derived, false when it is not

Devito.name — Method

name(subdomain)

returns the name of the subdomain

Devito.nsimplify — Method

nsimplify(expr::PyObject; constants=(), tolerance=none, full=false, rational=none, rational_conversion="base10")

Wrapper around sympy.nsimplify. Find a simple representation for a number or, if there are free symbols or if $rational=True$, then replace Floats with their Rational equivalents. If no change is made and rational is not False then Floats will at least be converted to Rationals. Explanation =========== For numerical expressions, a simple formula that numerically matches the given numerical expression is sought (and the input should be possible to evalf to a precision of at least 30 digits). Optionally, a list of (rationally independent) constants to include in the formula may be given. A lower tolerance may be set to find less exact matches. If no tolerance is given then the least precise value will set the tolerance (e.g. Floats default to 15 digits of precision, so would be tolerance=10**-15). With $full=True$, a more extensive search is performed (this is useful to find simpler numbers when the tolerance is set low). When converting to rational, if rationalconversion='base10' (the default), then convert floats to rationals using their base-10 (string) representation. When rationalconversion='exact' it uses the exact, base-2 representation.

See https://github.com/sympy/sympy/blob/52f606a503cea5e9588de14150ccb9f7f9ed4752/sympy/simplify/simplify.py .

Examples:

    nsimplify(π)

PyObject 314159265358979/100000000000000

    nsimplify(π; tolerance=0.1)

PyObject 22/7

Devito.origin — Method

origin(grid)

returns the tuple corresponding to the grid's origin

Devito.size_with_halo — Method

size_with_halo(x::DiscreteFunction)

Return the size of the grid associated with x, inclusive of the Devito "outer" halo.

Devito.size_with_inhalo — Method

size_with_inhalo(x::DiscreteFunction)

Return the size of the grid associated with z, inclusive the the Devito "inner" and "outer" halos.

Devito.solve — Method

solve(eq::PyObject, target::PyObject; kwargs...)

Algebraically rearrange an Eq w.r.t. a given symbol. This is a wrapper around $devito.solve$, which in turn is a wrapper around $sympy.solve$. Parameters ––––– eq : expr-like The equation to be rearranged. target : symbol The symbol w.r.t. which the equation is rearranged. May be a Function or any other symbolic object. **kwargs Symbolic optimizations applied while rearranging the equation. For more information. refer to $sympy.solve.__doc__$.

Devito.spacing — Function

spacing(x::Dimension)

Symbol representing the physical spacing along the Dimension.

Devito.stepping_dim — Method

stepping_dim(x::Grid)

Returns the stepping dimension for the associated grid.

Devito.subdomains — Method

subdomains(grid)

returns subdomains associated with a Devito grid

Devito.symbolic_max — Function

symbolic_max(x::Dimension)

Symbol defining the maximum point of the Dimension

Devito.symbolic_min — Function

symbolic_min(x::Dimension)

Symbol defining the minimum point of the Dimension

Devito.symbolic_size — Function

symbolic_size(x::Dimension)

Symbol defining the size of the Dimension

Devito.time_dim — Method

time_dim(x::Union{Grid,TimeFunction})

Returns the time dimension for the associated object.

Devito.value! — Method

value!(x::Constant{T},y::T)

Change the numerical value of a constant, x, after creation to y, after converting y to datatype T of constant x.

Devito.value — Method

value(x::Constant)

Returns the value of a devito constant. Can not be used to change constant value, for that use set_constant!(x,y)

Devito.Constant — Method

Constant(args...; kwargs...)

Symbol representing a constant, scalar value in symbolic equations. A Constant carries a scalar value. kwargs: name (str) – Name of the symbol. value (Real) – Value associated with the symbol. Defaults to 0. dtype (data-type, optional) – Any object that can be interpreted as a float data type. Defaults to Float32.

Devito.Function — Method

Devito.Function(; kwargs...)

Tensor symbol representing a discrete function in symbolic equations.

See: https://www.devitoproject.org/devito/function.html?highlight=function#devito.types.Function

Example

x = SpaceDimension(name="x", spacing=Constant(name="h_x", value=5.0))
z = SpaceDimension(name="z", spacing=Constant(name="h_z", value=5.0))
grid = Grid(
    dimensions = (x,z),
    shape = (251,501), # assume x is first, z is second (i.e. z is fast in python)
    origin = (0.0,0.0),
    extent = (1250.0,2500.0),
    dtype = Float32)

b = Devito.Function(name="b", grid=grid, space_order=8)

Devito.Grid — Method

Grid(; shape[, optional key-word arguments...])

Construct a grid that can be used in the construction of a Devito Functions.

See: https://www.devitoproject.org/devito/grid.html?highlight=grid#devito.types.Grid

Example

x = SpaceDimension(name="x", spacing=Constant(name="h_x", value=5.0))
z = SpaceDimension(name="z", spacing=Constant(name="h_z", value=5.0))
grid = Grid(
    dimensions = (x,z), # z is fast (row-major)
    shape = (251,501),
    origin = (0.0,0.0),
    extent = (1250.0,2500.0),
    dtype = Float32)

Devito.SparseTimeFunction — Method

SparseTimeFunction(; kwargs...)

Tensor symbol representing a space- and time-varying sparse array in symbolic equations.

See: https://www.devitoproject.org/devito/sparsetimefunction.html?highlight=sparsetimefunction

Example

x = SpaceDimension(name="x", spacing=Constant(name="h_x", value=5.0))
z = SpaceDimension(name="z", spacing=Constant(name="h_z", value=5.0))
grid = Grid(
    dimensions = (x,z),
    shape = (251,501), # assume x is first, z is second (i.e. z is fast in python)
    origin = (0.0,0.0),
    extent = (1250.0,2500.0),
    dtype = Float32)

time_range = 0.0f0:0.5f0:1000.0f0
src = SparseTimeFunction(name="src", grid=grid, npoint=1, nt=length(time_range))

Devito.SubDomain — Method

SubDomain(name, instructions)

Create a subdomain by passing a list of instructions for each dimension. Using an instruction with (nothing,) implies that the whole dimension should be used for that subdomain, as will ("middle",0,0)

Examples:

instructions = ("left",2),("middle",3,3)
SubDomain("subdomain_name",instructions)

or

instructions = [("right",4),("middle",1,2)]
SubDomain("subdomain_name",instructions)

or

SubDomain("subdomain_name",("right",2),("left",1))

Devito.TimeFunction — Method

TimeFunction(; kwargs...)

Tensor symbol representing a discrete function in symbolic equations.

See https://www.devitoproject.org/devito/timefunction.html?highlight=timefunction#devito.types.TimeFunction.

Example

x = SpaceDimension(name="x", spacing=Constant(name="h_x", value=5.0))
z = SpaceDimension(name="z", spacing=Constant(name="h_z", value=5.0))
grid = Grid(
    dimensions = (x,z),
    shape = (251,501), # assume x is first, z is second (i.e. z is fast in python)
    origin = (0.0,0.0),
    extent = (1250.0,2500.0),
    dtype = Float32)

p = TimeFunction(name="p", grid=grid, time_order=2, space_order=8)